test(reservation): ReservationDetails에 대한 테스트 작성 및 Reservation 테스트 메서드 수정
